pyproject.toml
[project]
name = "deepvuln"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.8.0",
    "jinja2==3.1.6",
    "json-repair==0.54.2",
    "langchain==1.1.2",
    "langchain-chroma==1.0.0",
    "langchain-community==0.4.1",
    "langchain-dev-utils[standard]>=1.2.16",
    "langchain-experimental>=0.4.0",
    "langchain-milvus==0.3.1",
    "langchain-openai==1.1.0",
    "langchain-qdrant==1.1.0",
    "langgraph==1.0.4",
    "langgraph-api==0.4.28",
    "langgraph-cli[inmem]>=0.4.7",
    "litellm>=1.80.5",
    "openai==2.9.0",
    "pydantic==2.12.5",
    "pymilvus==2.6.5",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "qdrant-client==1.16.1",
    "requests==2.32.5",
]
